Remove duplicates, remove useless Infernals
This commit removes the duplicate Infernal skeleton.
It also removes the mob spawn drop chance for very very basic Infernals that might be abused in the future to be grinded. 
We don't want that.
Also fixes Mmadrid's poor identation & too high spawn rates for basic mobs. (The Golem, what a bad name)